requirements.txt
anyio==3.7.1
appnope==0.1.3
asttokens==2.2.1
backcall==0.2.0
bcrypt==4.0.0
click==8.1.6
comm==0.1.4
debugpy==1.6.8
decorator==5.1.1
dnspython==2.4.1
email-validator==1.3.0
exceptiongroup==1.1.2
executing==1.2.0
fastapi==0.85.0
h11==0.14.0
idna==3.4
ipykernel==6.25.0
ipython==8.14.0
jedi==0.19.0
jupyter_client==8.3.0
jupyter_core==5.3.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.7
packaging==23.1
parso==0.8.3
passlib==1.7.4
pexpect==4.8.0
pickleshare==0.7.5
platformdirs==3.10.0
prompt-toolkit==3.0.39
psutil==5.9.5
psycopg2-binary==2.9.3
ptyprocess==0.7.0
pure-eval==0.2.2
pydantic==1.10.12
Pygments==2.15.1
python-dateutil==2.8.2
python-dotenv==1.0.0
python-multipart==0.0.6
pyzmq==25.1.0
six==1.16.0
sniffio==1.3.0
stack-data==0.6.2
starlette==0.20.4
tornado==6.3.2
traitlets==5.9.0
typing_extensions==4.7.1
uvicorn==0.23.2
wcwidth==0.2.6
